requirements.txt
bitsandbytes
cupy_cuda12x
diffusers
einops
geometry
huggingface_hub
transformers
accelerate
ipywidgets
joblib
k3d
matplotlib
numba
numpy
Pillow
scipy
torch
torch_optimizer
torchvision
tqdm
vtk
jupyter
scikit-sparse
datasets